requirements.txt
altgraph==0.17
macholib==1.14
Pillow==7.0.0
PyInstaller==3.6
